+++ src/backend/core/external_api/viewsets.py
@@ -158,6 +158,8 @@ class ApplicationViewSet(viewsets.ViewSet):
 class RoomViewSet(
     mixins.CreateModelMixin,
     mixins.RetrieveModelMixin,
+    mixins.UpdateModelMixin,
+    mixins.DestroyModelMixin,
     mixins.ListModelMixin,
     viewsets.GenericViewSet,
 ):